requirements.txt
numpy
scipy
keras
Pillow
mss
h5py
tensorflow
pytesseract
sklearn
matplotlib
